+++ src/components/Layouts/index.js
@@ -1,3 +1,4 @@
+import Carousel from "./Carousel";
 import GameLayout from "./GameLayout";
 
-export { GameLayout };
+export { GameLayout, Carousel };
