+++ src/components/Modals/index.js
@@ -1,5 +1,15 @@
 import ButtonModal from "./ButtonModal";
 import ContentModal from "./ContentModal";
+import FailModalContainer from "./FailModalContainer";
 import InputModalContainer from "./InputModalContainer";
+import NextStageModalContainer from "./NextStageModalContainer";
+import ResetModalContainer from "./ResetModalContainer";
 
-export { InputModalContainer, ContentModal, ButtonModal };
+export {
+  InputModalContainer,
+  ContentModal,
+  ButtonModal,
+  FailModalContainer,
+  NextStageModalContainer,
+  ResetModalContainer,
+};
