+++ src/components/index.js
@@ -3,7 +3,14 @@ import ChanceIcons from "./ChanceIcons";
 import { GameHeader } from "./Headers";
 import { GameLayout } from "./Layouts";
 import Logo from "./Logo";
-import { InputModalContainer, ContentModal, ButtonModal } from "./Modals";
+import {
+  InputModalContainer,
+  ContentModal,
+  ButtonModal,
+  NextStageModalContainer,
+  ResetModalContainer,
+  FailModalContainer,
+} from "./Modals";
 import { TextTimer } from "./Timers";
 
 export {
@@ -14,6 +21,9 @@ export {
   InputModalContainer,
   ContentModal,
   ButtonModal,
+  NextStageModalContainer,
+  ResetModalContainer,
+  FailModalContainer,
   GameHeader,
   GameLayout,
   TextTimer,
