+++ src/features/main/index.js
@@ -1,4 +1,5 @@
 import MainContainer from "./container/MainContainer";
 import Main from "./presenter/Main";
+import Page from "./presenter/Page";
 
-export { MainContainer, Main };
+export { MainContainer, Main, Page };
